pyproject.toml
[project]
name = "glomerulimonorepo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "PySide6==6.9.0",
    "openslide-python==1.4.2",
    "openslide-bin==4.0.0.8",
    "pillow==11.2.1",
    "torch==2.6.0",
    "torchvision==0.21.0",
    "scikit-learn==1.6.1",
    "opencv-python==4.11.0.86",
    "tzdata==2025.2",
]
